+++ core/compat/psp/syspatch.c
@@ -54,7 +54,7 @@ static int _sceKernelBootFromForUmdMan(void)
     return 0x20;
 }
 
-void patch_sceUmdMan_driver(SceModule* mod)
+void patch_sceUmdMan_driver(SceModule2* mod)
 {
     int apitype = sceKernelInitApitype();
     if (apitype == 0x152 || apitype == 0x141) {
@@ -174,7 +174,7 @@ int pause_disabled = 0;
 void disable_PauseGame()
 {
     if(psp_model == PSP_GO && !pause_disabled) {
-        SceModule2* mod = sceKernelFindModuleByName("sceImpose_Driver");
+        SceModule2* mod = (SceModule2*)sceKernelFindModuleByName("sceImpose_Driver");
         u32 text_addr = mod->text_addr;
         for(int i=0; i<2; i++) {
             _sw(NOP, text_addr + 0x00000574 + i * 4);
@@ -296,15 +296,6 @@ void PSPOnModuleStart(SceModule2 * mod){
         goto flush;
     }
 
-    if (strcmp(mod->modname, "MacroFire") == 0){
-        // fix for MacroFire (disables sceUmdActivate/Deactivate functions)
-        // this is needed because ARK loads plugins when UMD is already active (MediaSync fully loaded and started)
-        // while older CFW load plugins a bit earlier (MediaSync loaded but not started)
-        hookImportByNID(mod, "sceUmdUser", 0xC6183D47, 0);
-        hookImportByNID(mod, "sceUmdUser", 0xE83742BA, 0);
-        goto flush;
-    }
-
     if (strcmp(mod->modname, "DayViewer_User") == 0){
         // fix scePaf imports in DayViewer
         static u32 nids[] = {
@@ -356,6 +347,7 @@ void PSPOnModuleStart(SceModule2 * mod){
             switch (se_config->clock){
                 case 1: SetSpeed(333, 166); break;
                 case 2: SetSpeed(133, 66); break;
+                case 3: SetSpeed(222, 111); break;
             }
             // Boot Complete Action done
             booted = 1;