+++ core/vshctrl/vshmenu.c
@@ -38,10 +38,7 @@
 #define ALL_CTRL     (ALL_ALLOW|ALL_BUTTON|ALL_TRIGGER|ALL_FUNCTION)
 
 extern ARKConfig* ark_config;
-extern int cur_usbdevice;
-extern int usb_readonly;
-
-SEConfig conf;
+extern SEConfig* se_config;
 
 static int (*g_VshMenuCtrl) (SceCtrlData *, int);
 static SceUID g_satelite_mod_id = -1;
@@ -61,33 +58,20 @@ int vctrlVSHRegisterVshMenu(int (*ctrl)(SceCtrlData *, int))
 
 int vctrlVSHUpdateConfig(SEConfig *config)
 {
-    u32 k1;
-    int ret;
-    k1 = pspSdkSetK1(0);
-    memcpy(&conf, config, sizeof(conf));
-    ret = sctrlSESetConfig(&conf);
-    cur_usbdevice = config->usbdevice;
-    usb_readonly = config->usbdevice_rdonly;
-    pspSdkSetK1(k1);
-    return ret;
+    memcpy(se_config, config, sizeof(SEConfig));
+    return 0;
 }
 
 int vctrlVSHExitVSHMenu(SEConfig *config, char *videoiso, int disctype)
 {
-    u32 k1;
-    int ret = 0;
 
-    k1 = pspSdkSetK1(0);
     if (config){
-        ret = vctrlVSHUpdateConfig(config);
-        cur_usbdevice = config->usbdevice;
-        usb_readonly = config->usbdevice_rdonly;
+        vctrlVSHUpdateConfig(config);
     }
 
     g_VshMenuCtrl = NULL;
-    pspSdkSetK1(k1);
     
-    return ret;
+    return 0;
 }
 
 static SceUID load_satelite(void)